chore(routers): move /metadata from auth.py to root.py
