+++ app/routers/auth.py
@@ -2,27 +2,12 @@ import httpx
 from fastapi import APIRouter, Depends, HTTPException, Request
 from fastapi.responses import RedirectResponse
 from fastapi.security import HTTPAuthorizationCredentials, HTTPBearer
-from app.model import MetadataResponse
 from app.utils.auth_helpers import cache_cognito_metadata, generate_cognito_logout_url, redirect_to_cognito_login
 from app.utils.jwt_helpers import verify_and_extract_sub
 from app.utils.token_helpers import exchange_token
 
 router = APIRouter()
 bearer_scheme = HTTPBearer()
-
-@router.get('/metadata', response_model=MetadataResponse)
-def get_metadata():
-    return {
-        'login_endpoint': '/login',
-        'logout_endpoint': '/logout',
-        'verify_access_token_endpoint': '/token',
-        'verify_userinfo_endpoint': '/user',
-        'health_check_endpoint': '/health',
-        'simulate_404_endpoint': '/error/404',
-        'docs_url': '/docs',
-        'redoc_url': '/redoc',
-        'openapi_url': '/openapi.json',
-    }
 
 @router.get('/login')
 async def login(request: Request) -> RedirectResponse:
